% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soccer_functions.R
\name{assists_per_game}
\alias{assists_per_game}
\title{Calculate Assists Per Game}
\usage{
assists_per_game(assists, games_played)
}
\arguments{
\item{assists}{A numeric value. The total number of assists by the player.}

\item{games_played}{A numeric value. The total number of games played by the player.}
}
\value{
A numeric value representing the average assists per game.
}
\description{
This function calculates the average assists per game for a soccer player.
}
\examples{
# Example 1: Normal case
assists_per_game(10, 25)

}
